added sample data to trial account

+++ src/app/api/jobs/delete/route.tsx
@@ -12,6 +12,7 @@ export async function DELETE(request: NextRequest) {
       let db_connect = client.db("jobsData");
       const { searchParams } = new URL(request.url!);
       let user_id: string  = searchParams.get("id") || "";
+      if (user_id === "6482c564b18df6bd4874cb5c") return NextResponse.json({message: "Cannot use the trial user_id"})
       let jobId: string  = searchParams.get("jobid") || "";
       let query = { _id: new ObjectId(jobId), user_id: user_id };
       let deleteResult = await db_connect.collection("jobsData").deleteOne(query);
